+++ projects/ngx-testing-tools/src/lib/common/test-bed/base/base-test-bed-factory.ts
@@ -1,9 +1,9 @@
 import { ProviderToken, Type } from '@angular/core';
 import { TestBed, TestBedStatic } from '@angular/core/testing';
-import { AnyProvider, Importation } from '../../../component/test-bed/models/metadata-type.model';
 import { MaybeArray, NonEmptyString, PrettyMerge } from '../../../shared.model';
 import { makeArray } from '../../../util/array.util';
 import { EnhancedJasmineCallback } from '../models/enhanced-jasmine-callback.model';
+import { AnyProvider, Importation } from '../models/metadata-type.model';
 import { InjectionStore } from '../store/models/injected-store.model';
 import { BaseTestBedOptions } from './models/base-test-bed-options.model';
