+++ projects/ngx-testing-tools/src/lib/common/test-bed/base/models/base-test-bed-options.model.ts
@@ -1,4 +1,4 @@
-import { AnyProvider, Importation } from '../../../../component/test-bed/models/metadata-type.model';
+import { AnyProvider, Importation } from '../../models/metadata-type.model';
 
 export interface BaseTestBedOptions {
   /**
