+++ projects/ngx-testing-tools/src/tests/fixtures/mocks/test-compiler.mock.ts
@@ -1,5 +1,5 @@
 import { TestModuleMetadata } from '@angular/core/testing';
-import { AnyProvider, Declaration, Importation } from '../../../lib/component/test-bed/models/metadata-type.model';
+import { AnyProvider, Declaration, Importation } from '../../../lib/common/test-bed/models/metadata-type.model';
 
 export class MockTestCompiler {
